+++ bz-webserver/attributes/nginx.rb
@@ -1,0 +1,2 @@
+# opens port 80 for rhel
+default['bz-webserver']['open-80-port'] = true